Refactor: Add a new VehiclePositionsPipeline
Split this functionality out from Concentrate.Supervisor.
